Added day 9
Not 100% sure about the solution to part 2, but it gives the right answer.
Rather than using the local minima to calculate the size of the basins, I just segmented the height map between the 9 height points.

It isn't quite the same as the problem described, but I believe any input that wouldn't work for this solution also would have ambiguous points between basins. The problem states that only points with height 9 don't belong to a basin and all others belong to exactly one basin. So this solution should work with any inputs that would be valid for the stated problem (and obviously it gave the right solution for the input data).